refactor: :recycle: src/components/layouts 폴더 추가
