+++ src/components/LevelModal/LevelModal.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useState } from 'react';
 
 import { Level as LevelList } from '@/constants/Level';
-import ModalLayout from '@/components/ModalLayout';
+import ModalLayout from '@/components/layouts/ModalLayout';
 import Icon from '@/components/commons/Icon';
 import { theme } from '@/themes';
