+++ src/components/Modal/Modal.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import React, { ReactNode } from 'react';
 
-import ModalLayout from '@/components/ModalLayout';
+import ModalLayout from '@/components/layouts/ModalLayout';
 import Icon from '@/components/commons/Icon';
 
 interface ModalProps {
