+++ src/containers/CreateRecordContainer/RecordSecondStepContainer/RecordSecondStepContainer.tsx
@@ -12,7 +12,7 @@ import BottomFloatingButtonArea from '@/components/BottomFloatingButtonArea';
 import Button from '@/components/commons/Button';
 import FormSubmitButton from '@/components/commons/FormSubmitButton';
 import Icon from '@/components/commons/Icon';
-import { SwiperLayoutChildProps } from '@/components/SwiperLayout';
+import { SwiperLayoutChildProps } from '@/components/layouts/SwiperLayout';
 
 const StyledRecordSecondStepContainer = styled.div`
   display: flex;
