+++ src/containers/CreateRecordContainer/RecordThirdStepContainer/RecordThirdStepContainer.tsx
@@ -13,8 +13,8 @@ import SelectField from '@/components/formFields/SelectField';
 import TextAreaField from '@/components/formFields/TextAreaField';
 import BottomFloatingButtonArea from '@/components/BottomFloatingButtonArea';
 import Button, { ButtonCount } from '@/components/commons/Button';
-import FormSubmitButton from '@/components/commons/FormSubmitButton';
-import { SwiperLayoutChildProps } from '@/components/SwiperLayout';
+import FormSubmitButton from '@/components/layouts/SwiperLayout';
+import { SwiperLayoutChildProps } from '@/components/layouts/SwiperLayout';
 import Icon from '@/components/commons/Icon';
 
 interface RecordThirdStepContainerProps extends SwiperLayoutChildProps {
